requirements.txt
streamlit
pandas
gspread
google-auth-oauthlib
google-auth
google-auth-httplib2
requests
python-dotenv
openai
